move products state to ProductsPage
